requirements.txt
streamlit
groq
requests
PyPDF2
python-docx
sentence-transformers
numpy
python-dotenv
